requirements.txt
pandas
groq
scikit-learn
numpy
django
django-cors-headers
djangorestframework
django-filter
python-dotenv
ipykernel
pyyaml
python-dotenv
openpyxl
requests
matplotlib
seaborn
markdown
kneed